Refactor CustomerClient to use pydantic types
